requirements.txt
flask
black
flask-restful
flask-sqlalchemy
marshmallow
flask-jwt-extended
python-dotenv
torch
torchvision